fix and draw function

+++ src/structures/classes/buttons/connect/redirect.connect.js
@@ -20,15 +20,16 @@ export default ({ interaction, user, message }, commandData) => {
 
     function cancel() {
 
-        if (userId !== user.id)
+        if (![userId, message.interaction.user.id].includes(user.id))
             return interaction.reply({
                 content: `${e.DenyX} | Eeepa, você não usou o comando e também não foi desafiado, né?`,
                 ephemeral: true
             })
 
+        Database.Cache.Connect.delete(message.id)
         return interaction.update({
-            content: `${e.Deny} | <@${userId}> cancelou o desafio.`,
-            ephemeral: true
+            content: `${e.Deny} | ${interaction.user} cancelou o desafio Connect4.`,
+            components: []
         }).catch(() => { })
     }
 
@@ -105,6 +106,7 @@ export default ({ interaction, user, message }, commandData) => {
             ephemeral: true
         })
             .catch(err => {
+                if (err.code == 10062) return interaction.channel.send({ connect: `${e.cry} | ${interaction.user}, o Discord não entregou todos os dados necessário. Pode clicar no botão mais uma vez?` })
                 Database.Cache.Connect.delete(message.id)
                 message.delete().catch(() => { })
                 return interaction.channel.send({ content: `${e.cry} | Erro ao iniciar o jogo\n${e.bug} | \`${err}\`` })
